PatientDeskAgent/medication_agent/requirements.txt
google-adk[a2a]==1.18.0
pandas==2.3.3
fhirclient>=4.3.2